refactor: move DocumentLinkAdapter to new file
